% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gN3dist.R
\docType{data}
\name{gN3dist}
\alias{gN3dist}
\title{Distance between the centroids of NUTS3 regions in Germany}
\format{
matrix of distances
}
\usage{
gN3dist
}
\description{
Distance between the centroids of NUTS3 regions in Germany
}
\keyword{data}
